LPS-148749 search-experiences-web: Move, rename, and update schema
